backend/requirements.txt
Flask==2.3.2
requests
pandas
openpyxl
python-dotenv
pyjwt
flask-cors
